requirements.txt
alembic==1.0.0
APScheduler==3.5.3
blinker==1.4
certifi==2018.10.15
chardet==3.0.4
click==6.7
dominate==2.3.4
Flask==1.0.2
Flask-APScheduler==1.10.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
nltk==3.3
passlib==1.7.1
psutil==5.4.8
psycopg2==2.7.6.1
PyJWT==1.6.4
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
requests==2.20.0
six==1.11.0
speaklater==1.3
SQLAlchemy==1.2.11
tzlocal==1.5.1
urllib3==1.24
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
WTForms-JSON==0.3.3
